Handle :refer'd macros more gracefully
Clojurescript also allows you to simply :refer macros as sugar for
:require-macros now, but the ns parsing (as expected) doesn't *know*
that it's a macro. Future work might be able to detect that it's a
macro and a valid var, but right now we just ignore it; I'd thought
about just assuming it's a macro, but that kind of gives a visual
indication suggesting it's valid when we can't confirm it is, which
doesn't seem ideal.